refactor(mlir): streamline C++ header collection and remove unused files
- Replaced the `_collect_pch_headers_from_device_cpp` function with inline logic in `_cmd_build` to enhance clarity and reduce function overhead.
- Removed the `CppPlacement.h` header file and its associated implementation, as it was no longer needed after refactoring.
- Updated `CMakeLists.txt` to remove references to the deleted `CppPlacement.cpp` file, ensuring a cleaner build configuration.
- Enhanced `CppEmitter.h` with new placement attributes and utility functions to support the updated logic.

These changes improve the maintainability and efficiency of the C++ emitter's header management.
