+++ compiler/mlir/CMakeLists.txt
@@ -87,7 +87,6 @@ add_library(pyc_transforms
   lib/Transforms/PackI1RegsPass.cpp
   lib/Transforms/FuseCombPass.cpp
   lib/Transforms/CppPlacementPass.cpp
-  lib/Emit/CppPlacement.cpp
   lib/Transforms/LowerSCFToPYCStaticPass.cpp
   lib/Transforms/PrunePortsPass.cpp
   lib/Transforms/SLPPackWiresPass.cpp
